+++ src/Sunbay.Nexus.Sdk/Models/Common/RefundAmount.cs
@@ -5,39 +5,40 @@ namespace Sunbay.Nexus.Sdk.Models.Common
     /// <summary>
     /// Refund amount information
     /// Supports: orderAmount, tipAmount, taxAmount, surchargeAmount, cashbackAmount
+    /// All amounts are in the smallest currency unit (e.g., cents for USD, fen for CNY)
     /// </summary>
     public class RefundAmount
     {
         /// <summary>
-        /// Order amount (required)
+        /// Order amount in smallest currency unit (required)
         /// </summary>
         [JsonPropertyName("orderAmount")]
-        public decimal OrderAmount { get; set; }
+        public long OrderAmount { get; set; }
         
         /// <summary>
-        /// Tip amount (optional, must be greater than or equal to 0)
+        /// Tip amount in smallest currency unit (optional, must be greater than or equal to 0)
         /// </summary>
         [JsonPropertyName("tipAmount")]
-        public decimal? TipAmount { get; set; }
+        public long? TipAmount { get; set; }
         
         /// <summary>
-        /// Tax amount (optional, must be greater than or equal to 0)
+        /// Tax amount in smallest currency unit (optional, must be greater than or equal to 0)
         /// </summary>
         [JsonPropertyName("taxAmount")]
-        public decimal? TaxAmount { get; set; }
+        public long? TaxAmount { get; set; }
         
         /// <summary>
-        /// Surcharge amount (optional, must be greater than or equal to 0).
+        /// Surcharge amount in smallest currency unit (optional, must be greater than or equal to 0).
         /// Note: Some processors may require surcharge to be refunded proportionally. Please contact technical support for detailed policies.
         /// </summary>
         [JsonPropertyName("surchargeAmount")]
-        public decimal? SurchargeAmount { get; set; }
+        public long? SurchargeAmount { get; set; }
         
         /// <summary>
-        /// Cashback amount (optional, must be greater than or equal to 0)
+        /// Cashback amount in smallest currency unit (optional, must be greater than or equal to 0)
         /// </summary>
         [JsonPropertyName("cashbackAmount")]
-        public decimal? CashbackAmount { get; set; }
+        public long? CashbackAmount { get; set; }
         
         /// <summary>
         /// Pricing currency (ISO 4217, required)
